tracking.py: include frames without detections in the tracking phase (#2)
Frames without detections were being dropped in the detect phase which
means they never reached the tracking step and so were not included in
the results VIA project.

I'm not sure why the frames were being dropped but one effect is what
happens to the track ids.  If detection fails only in a single frame,
then the track is effectively not broken.  Including all frames, will
lead to more tracks found (which can be argued is also more correct).