track: use tracker to avoid track interruptions (#10)
svt.detections.detections.match() only uses the tracker to select
boxes from the detection step.  If the detector fails in some frames
it will not fill the track with the tracker results.

This commit adds a svt_patch that does it.  We probably should add new
methods to the svt package.

This also lowers the default MATCH_OVERLAP_THRESHOLD which we found
too conservative.