Use detectron2 PathManager to download remote config and model files.
Detectron2 config files and svt and detectron2 weight files can be
specified as URLs.  We downloaded the fike ourselves.  It didn't
delete the temporary files and always downloaded a new one.  Because
the file names were random it was difficult to know where it came from
the logs and it missed th file extension that some functions that
would read them relied on.  It was also duplicated between the
tracking.ipynb and __main__.py.  These were all fixable but we can
just use detectron2 PathManager.