detect: do not filter detections by confidence score.
We want to be able to filter detections by confidence score but we
also want to export all detections, even those below the threshold, to
a VIA project for debugging.  Also, we later will want to reuse
previous detections results for tracking with different threshold
values.  So we can't perform the filtering inside `detect()`.